+++ CHANGELOG.md
@@ -27,6 +27,7 @@ Bug fixes:
 * [OLMIS-3727](https://openlmis.atlassian.net/browse/OLMIS-3727): Fixed getting user's programs.
 * [OLMIS-3468](https://openlmis.atlassian.net/browse/OLMIS-3468): Fixed a bug with missing permissions after a relog.
 * [OLMIS-3806](https://openlmis.atlassian.net/browse/OLMIS-3806): Fixed date picker to not change selected date
+* [OLMIS-3759](https://openlmis.atlassian.net/browse/OLMIS-3759): Fixed program facility select to disable "supervised facility" if none to supervise.
 
 
 5.2.2 / 2017-11-09